+++ HISTORY.rst
@@ -3,6 +3,10 @@
 History
 -------
 
+v0.x.x (xxx-xx-xx)
+..................
+* fix schema for ``timedelta`` as number, by @tiangolo
+
 v0.16.1 (2018-12-10)
 ....................
 * fix ``create_model`` to correctly use the passed ``__config__``, #320 by @hugoduncan